`EmberEntropySource`
